lib/core/pyproject.toml
[project]
name = "qbrixcore"
version = "0.1.0alpha0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.2.1",
    "numpy>=2.2.6",
    "pydantic-settings>=2.9.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pre-commit>=4.0.1",
    "black>=24.0.0",
    "pytest-mock>=3.14.1",
]

[[tool.uv.index]]
name = "envelope"
url = "https://app.envelope.dev/legacy/"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=qbrixcore",
    "--cov-report=term-missing",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "performance: Performance tests",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"